Refactor devtool panel components
